Removed old file Stock which was used for the oop approach fixed errors in trade.py related to try except blocks did some testing for writing to json files in the apitest file
changed the implementation for the trading bot to use dictionarys
rather than the oop approach i was taking beforehand. this is done
to make writing to a json file simpler as well as removing unneeded
complexity from the program that was introduced by the oop approach
all the data is now being stored in dictionarys which can easily be
written into a json file.

For now all of the relevant methods have been implemented
but not yet tested.
all of the methods that interact with the api have a loop to force them
to try and connect to the api until it answers with a request.
furthermore there are also checks in place to avoid crashes related
to the fact that sometimes the stocks of interest cant be found
trough the alpha vantage api

all of the methods are simply called from one main loop for the programm.

Things that have to still be worked on are the point at which a stock
is supposed to be sold, as well as keeping track of possibly more
data related to the value of any given stock